+++ src/main/java/calculator/Application.java
@@ -1,5 +1,8 @@
 package calculator;
 
+import calculator.common.CalculatorAppConfig;
+import calculator.controller.CalculatorApplication;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
