+++ src/main/java/calculator/common/ConsoleMessage.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.common;
 
 public class ConsoleMessage {
