+++ src/main/java/calculator/view/ApplicationView.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view;
 
 public interface ApplicationView {
