+++ src/main/java/calculator/view/input/InputView.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view.input;
 
 public interface InputView {
