+++ src/main/java/calculator/view/output/OutputView.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view.output;
 
 public interface OutputView {
